implemented basic dashboard with navBar, FeaturedCard, SessionCards, and SessionsList
